refactor, linter
