+++ cmd/cli/pipeline.tmpl
