+++ cmd/cli/pipelines_init.go
